+++ src/main/java/chess/domain/game/ChessGame.java
@@ -1,24 +1,13 @@
 package chess.domain.game;
 
 import chess.domain.PiecesPosition;
-import chess.domain.piece.Camp;
 
-public abstract class ChessGame {
+public interface ChessGame {
 
-    protected final PiecesPosition piecesPosition;
-    protected final Camp turnCamp;
+    boolean isGameRunnable();
 
-    public ChessGame(PiecesPosition piecesPosition, Camp turnCamp) {
-        this.piecesPosition = piecesPosition;
-        this.turnCamp = turnCamp;
-    }
+    ChessGame playByCommand(ChessGameCommand gameCommand);
 
-    public abstract boolean isGameRunnable();
-
-    public abstract ChessGame playByCommand(ChessGameCommand gameCommand);
-
-    public PiecesPosition getPiecesPosition() {
-        return piecesPosition;
-    }
+    PiecesPosition getPiecesPosition();
 }
