+++ src/main/java/chess/domain/game/state/ReadyChessGame.java
@@ -1,15 +1,19 @@
-package chess.domain.game;
+package chess.domain.game.state;
 
 import chess.domain.PiecesPosition;
+import chess.domain.game.ChessGame;
+import chess.domain.game.ChessGameCommand;
 import chess.domain.piece.Camp;
 
-public class ReadyChessGame extends ChessGame {
+public class ReadyChessGame implements ChessGame {
 
     private static final String NOT_START_COMMAND = "게임이 시작되지 않았습니다. 게임을 먼저 시작해주세요.";
-    private static final Camp FIRST_CAMP = Camp.WHITE;
+    private static final Camp INIT_CAMP = Camp.WHITE;
+
+    private final PiecesPosition piecesPosition;
 
     public ReadyChessGame(PiecesPosition piecesPosition) {
-        super(piecesPosition, FIRST_CAMP);
+        this.piecesPosition = piecesPosition;
     }
 
     @Override
@@ -20,9 +24,14 @@ public class ReadyChessGame extends ChessGame {
     @Override
     public ChessGame playByCommand(ChessGameCommand gameCommand) {
         if (gameCommand.isStart()) {
-            return new RunningChessGame(getPiecesPosition(), FIRST_CAMP);
+            return new RunningChessGame(piecesPosition, INIT_CAMP);
         }
 
         throw new IllegalStateException(NOT_START_COMMAND);
+    }
+
+    @Override
+    public PiecesPosition getPiecesPosition() {
+        return this.piecesPosition;
     }
 }
