+++ src/test/java/chess/domain/game/state/RunningChessGameTest.java
@@ -1,10 +1,13 @@
-package chess.domain.game;
+package chess.domain.game.state;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import chess.domain.PiecesPosition;
+import chess.domain.game.ChessCommandType;
+import chess.domain.game.ChessGame;
+import chess.domain.game.ChessGameCommand;
 import chess.domain.piece.Camp;
 import chess.domain.position.File;
 import chess.domain.position.Position;
